# api/ewbi/swagger.yaml
...
openapi: 3.0.3
info:
  version: 1.4.0
  title: Federation Management Service
  description: |
    # Introduction
    ---
    RESTful APIs that allow an OP to share the edge cloud resources and capabilities securely to other partner OPs over E/WBI.

    ---
    # API Scope

    ---
    APIs defined in this version of the specification can be categorized into the following areas:
    * __FederationManagement__ - Create and manage directed federation relationship with a partner OP
    * __AvailabilityZoneInfoSynchronization__ - Management of resources of partner OP zones and status updates
    * __ArtefactManagement__ - Upload, remove, retrieve and update application descriptors, charts and packages over E/WBI towards a partner OP
    * __FileManagement__ - Upload, remove, retrieve and update application binaries over E/WBI towards a partner OP
    * __ApplicationOnboardingManagement__ - Register, retrieve, update and remove applications over E/WBI towards a partner OP
    * __ApplicationDeploymentManagement__ - Create, update, retrieve and terminate application instances over E/WBI towards a partner OP
    * __AppProviderResourceManagement__ - Static resource reservation for an application provider over E/WBI for partner OP zones
    * __EdgeNodeSharing__ - Edge discovery procedures towards partner OP over E/WBI.
    * __LBORoamingAuthentication__ - Validation of user client authentication from home OP

    ---
    # Definitions
    ---
...

...
    * __OP__ - Operator platform.
...

...

...

...

...

...

...
components:
  schemas:
...
    FederationContextId:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9-]*$
      readOnly: true
      description: This identifier shall be provided by the partner OP on successful verification and validation of the federation create request and is used by partner op to identify this newly created federation context. Originating OP shall provide this identifier in any subsequent request towards the partner op.
...
    FederationDetailsCallbackRequestData:
      type: object
      required:
        - federationContextId
      properties:
        federationContextId:
          $ref: '#/components/schemas/FederationContextId'
...
  responses:
    "400":
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    "401":
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    "409":
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    "422":
      description: Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    default:
      description: Generic Error
paths:
...
  /{federationContextId}/partner:
    post:
      summary: Registers a callback to be called when there are updates on details about the federation context with the partner OP. The callback body shall provide info about the zones offered by the partner, partner OP network codes, information about edge discovery and LCM service etc.
      operationId: PartnerDetails
      tags:
        - FederationManagement
      parameters:
        - name: federationContextId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/FederationContextId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederationDetailsCallbackRequestData'
      responses:
        "200":
          description: Federation response data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FederationDetailResponseData'
        "400":
          $ref: '#/components/responses/400'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
      callbacks:
        onFederationDetailsChangeEvent:
          '{$request.body#/partnerDetailsCallbackLink}':
            post:
              requestBody:
                description: |
                  OP uses this callback api to notify partner OP about change in federation details
                content:
                  application/json:
                    schema:
                      type: object
                      properties:
                        edgeDiscoveryServiceEndPoint:
                          $ref: '#/components/schemas/ServiceEndpoint'
                        lcmServiceEndPoint:
                          $ref: '#/components/schemas/ServiceEndpoint'
                        allowedMobileNetworkIds:
                          $ref: '#/components/schemas/MobileNetworkIds'
                        allowedFixedNetworkIds:
                          $ref: '#/components/schemas/FixedNetworkIds'
                        offeredAvailabilityZones:
                          type: array
                          items:
                            $ref: '#/components/schemas/ZoneDetails'
                          minItems: 1
                        platformCaps:
                          $ref: '#/components/schemas/PlatformCaps'
              responses:
                "204":
                  description: Expected response to a successful call back processing
                "400":
                  $ref: '#/components/responses/400'
                "401":
                  $ref: '#/components/responses/401'
                "404":
                  $ref: '#/components/responses/404'
                "409":
                  $ref: '#/components/responses/409'
                "422":
                  $ref: '#/components/responses/422'
                "500":
                  $ref: '#/components/responses/500'
                "503":
                  $ref: '#/components/responses/503'
                "520":
                  $ref: '#/components/responses/520'
                default:
                  $ref: '#/components/responses/default'
              security:
                - notifClientCredentials: [fed-mgmt-notif]
    get:
      summary: Retrieves details about the federation context with the partner OP. The response shall provide info about the zones offered by the partner, partner OP network codes, information about edge discovery and LCM service etc.
      operationId: GetFederationDetails
      tags:
        - FederationManagement
      parameters:
        - name: federationContextId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/FederationContextId'
      responses:
        "200":
          description: Federation meta-info request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  edgeDiscoveryServiceEndPoint:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  lcmServiceEndPoint:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  allowedMobileNetworkIds:
                    $ref: '#/components/schemas/MobileNetworkIds'
                  allowedFixedNetworkIds:
                    $ref: '#/components/schemas/FixedNetworkIds'
                  offeredAvailabilityZones:
                    type: array
                    items:
                      $ref: '#/components/schemas/ZoneDetails'
                    minItems: 1
                  platformCaps:
                    $ref: '#/components/schemas/PlatformCaps'
        "400":
          $ref: '#/components/responses/400'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
    patch:
      summary: API used by the Originating OP towards the partner OP, to update the parameters associated to the existing federation
      operationId: UpdateFederation
      tags:
        - FederationManagement
      parameters:
        - name: federationContextId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/FederationContextId'
      requestBody:
        required: true
        description: Details about changes origination OP wished to apply
        content:
          application/json:
            schema:
              type: object
              required:
                - objectType
                - operationType
                - modificationDate
              properties:
                objectType:
                  type: string
                  enum:
                    - MOBILE_NETWORK_CODES
                    - FIXED_NETWORK_CODES
                    - OPS_POLICY
                    - APP_POLICY
                operationType:
                  type: string
                  enum:
                    - ADD_CODES
                    - REMOVE_CODES
                    - UPDATE_CODES
                    - ADD_POLICY
                    - REMOVE_POLICY
                    - UPDATE_POLICY
                addMobileNetworkIds:
                  $ref: '#/components/schemas/MobileNetworkIds'
                removeMobileNetworkIds:
                  $ref: '#/components/schemas/MobileNetworkIds'
                addFixedNetworkIds:
                  $ref: '#/components/schemas/FixedNetworkIds'
                removeFixedNetworkIds:
                  $ref: '#/components/schemas/FixedNetworkIds'
                assocAppPolicies:
                  $ref: '#/components/schemas/AssocApplPolicies'
                assocOpsPolicies:
                  $ref: '#/components/schemas/AssocOpsPolicies'
                modificationDate:
                  type: string
                  format: date-time
                  description: Date and time of the federation modification by the originating partner OP
      responses:
        "200":
          description: Federation meta-info request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  edgeDiscoveryServiceEndPoint:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  lcmServiceEndPoint:
                    $ref: '#/components/schemas/ServiceEndpoint'
                  allowedMobileNetworkIds:
                    $ref: '#/components/schemas/MobileNetworkIds'
                  allowedFixedNetworkIds:
                    $ref: '#/components/schemas/FixedNetworkIds'
                  offeredAvailabilityZones:
                    type: array
                    items:
                      $ref: '#/components/schemas/ZoneDetails'
                    minItems: 1
        "400":
          $ref: '#/components/responses/400'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
    delete:
      summary: Remove existing federation with the partner OP
      operationId: DeleteFederationDetails
      tags:
        - FederationManagement
      parameters:
        - name: federationContextId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/FederationContextId'
      responses:
        "200":
          description: Federation removed successfully
        "400":
          $ref: '#/components/responses/400'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
...
  /{federationCallbackId}/partnerStatusLink:
    post:
      operationId: PartnerStatusLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: |
          OP uses this callback api to notify partner OP about change in federation status, federation metadata or offered zone details. Allowed combinations of objectType and operationType are
          - FEDERATION - STATUS: Status specified by parameter 'federationStatus'.
          - ZONES - STATUS: Status specified by parameter 'zoneStatus'.
          - ZONES - ADD: Use parameter 'addZones' to define add new zones
          - ZONES - REMOVE: Use parameter 'removeZones' to define remove zones.
          - EDGE_DISCOVERY_SERVICE - UPDATE: Use parameter 'edgeDiscoverySvcEndPoint' to specify new endpoints
          - LCM_SERVICE - UPDATE: Use parameter 'lcmSvcEndPoint' to specify new endpoints
          - MOBILE_NETWORK_CODES - ADD: Use parameter 'addMobileNetworkIds' to define new mobile network codes.
          - MOBILE_NETWORK_CODES - REMOVE: Use parameter 'removeMobileNetworkIds' to remove mobile network codes.
          - FIXED_NETWORK_CODES - ADD: Use parameter 'addFixedNetworkIds' to define new fixed network codes.
          - FIXED_NETWORK_CODES - REMOVE: Use parameter 'removeFixedNetworkIds' to remove fixed network codes.
        content:
          application/json:
            schema:
              type: object
              required:
                - federationContextId
                - objectType
                - operationType
                - modificationDate
              properties:
                federationContextId:
                  $ref: "#/components/schemas/FederationContextId"
                objectType:
                  type: string
                  enum:
                    - FEDERATION
                    - ZONES
                    - EDGE_DISCOVERY_SERVICE
                    - LCM_SERVICE
                    - MOBILE_NETWORK_CODES
                    - FIXED_NETWORK_CODES
                operationType:
                  type: string
                  enum:
                    - STATUS
                    - UPDATE
                    - ADD
                    - REMOVE
                edgeDiscoverySvcEndPoint:
                  $ref: "#/components/schemas/ServiceEndpoint"
                lcmSvcEndPoint:
                  $ref: "#/components/schemas/ServiceEndpoint"
                addMobileNetworkIds:
                  $ref: "#/components/schemas/MobileNetworkIds"
                removeMobileNetworkIds:
                  $ref: "#/components/schemas/MobileNetworkIds"
                addFixedNetworkIds:
                  $ref: "#/components/schemas/FixedNetworkIds"
                removeFixedNetworkIds:
                  $ref: "#/components/schemas/FixedNetworkIds"
                addZones:
                  type: array
                  items:
                    $ref: "#/components/schemas/ZoneDetails"
                  description: List of zones, which the operator platform wishes to make available to developers/ISVs of requesting operator platform.
                  minItems: 1
                removeZones:
                  type: array
                  items:
                    $ref: "#/components/schemas/ZoneIdentifier"
                  description: List of zones, which the operator platform no longer wishes to share.
                  minItems: 1
                zoneStatus:
                  type: array
                  items:
                    type: object
                    required:
                      - zoneId
                      - status
                    properties:
                      zoneId:
                        $ref: "#/components/schemas/ZoneIdentifier"
                      status:
                        $ref: "#/components/schemas/Status"
                  minItems: 1
                federationStatus:
                  $ref: "#/components/schemas/Status"
                modificationDate:
                  type: string
                  format: date-time
                  description: Date and time of the federation modification by the originating partner OP
      responses:
        "204":
          description: Expected response to a successful call back processing
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/availZoneNotifLink:
    post:
      operationId: AvailZoneNotifLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification about resource availability.
        content:
          application/json:
            schema:
              type: object
              required:
                - federationContextId
                - zoneId
                - zoneResUpdInfo
              properties:
                federationContextId:
                  $ref: "#/components/schemas/FederationContextId"
                zoneId:
                  $ref: "#/components/schemas/ZoneIdentifier"
                zoneResUpdInfo:
                  type: array
                  items:
                    type: object
                    minProperties: 1
                    properties:
                      availableCompResources:
                        description: Resources exclusively reserved for the originator OP.
                        type: array
                        items:
                          $ref: "#/components/schemas/ComputeResourceInfo"
                        minItems: 1
                      availableNetResources:
                        type: object
                        properties:
                          egressBandWidth:
                            type: integer
                            format: int32
                            description: Max dl throughput that this edge can offer. It is defined in Mbps.
                          dedicatedNIC:
                            type: integer
                            format: int32
                          supportSriov:
                            type: boolean
                            description: If this zone support SRIOV networks or not
                          supportDPDK:
                            type: boolean
                            description: If this zone supports DPDK based networking
                        minProperties: 1
      responses:
        "200":
          description: Zone info notification acknowledged
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/appStatusCallbackLink:
    post:
      operationId: AppStatusCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - appId
                - statusInfo
              properties:
                appId:
                  $ref: "#/components/schemas/AppIdentifier"
                statusInfo:
                  type: array
                  items:
                    type: object
                    required:
                      - zoneId
                      - onboardStatusInfo
                    properties:
                      zoneId:
                        $ref: "#/components/schemas/ZoneIdentifier"
                      onboardStatusInfo:
                        description: Defines change in application status. This change could be related to application itself or an application instance status
                        type: string
                        enum:
                          - PENDING
                          - ONBOARDED
                          - DEBOARDING
                          - REMOVED
                          - FAILED
                  minItems: 1
      responses:
        "204":
          description: Application status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/artefactStatusCallbackLink:
    post:
      operationId: ArtefactStatusCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - artefactId
                - updateStatus
              properties:
                artefactId:
                  $ref: "#/components/schemas/ArtefactId"
                updateStatus:
                  description: Defines change in artefact status. This change could be related to the artefact itself or an artefact instance status
                  type: string
                  enum:
                    - PENDING
                    - ERROR
                    - READY
                    - UNKNOWN
      responses:
        "204":
          description: Artefact status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/fileStatusCallbackLink:
    post:
      operationId: FileStatusCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - fileId
                - updateStatus
              properties:
                fileId:
                  $ref: "#/components/schemas/FileId"
                updateStatus:
                  description: Defines change in file status. This change could be related to the file itself or a file instance status
                  type: string
                  enum:
                    - PENDING
                    - ERROR
                    - READY
                    - UNKNOWN
      responses:
        "204":
          description: File status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/appInstCallbackLink:
    post:
      operationId: AppInstCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - appId
                - appInstanceId
                - zoneId
                - appInstanceInfo
              properties:
                appId:
                  $ref: "#/components/schemas/AppIdentifier"
                appInstanceId:
                  $ref: "#/components/schemas/InstanceIdentifier"
                zoneId:
                  $ref: "#/components/schemas/ZoneIdentifier"
                appInstanceInfo:
                  type: object
                  properties:
                    appInstanceState:
                      $ref: "#/components/schemas/InstanceState"
                    message:
                      type: string
                      description: Event information or failure message.
                    accesspointInfo:
                      $ref: "#/components/schemas/AccessPointInfo"
                  minProperties: 1
                modificationDate:
                  type: string
                  format: date-time
                  description: Date and time of the instance state modification by partner OP.
      responses:
        "204":
          description: Application instance state notification acknowledged
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/resourceReservationCallbackLink:
    post:
      operationId: ResourceReservationCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - federationContextId
                - zoneId
                - appProviderId
                - poolId
                - grantedFlavours
              properties:
                federationContextId:
                  $ref: "#/components/schemas/FederationContextId"
                zoneId:
                  $ref: "#/components/schemas/ZoneIdentifier"
                appProviderId:
                  $ref: "#/components/schemas/AppProviderId"
                poolId:
                  $ref: "#/components/schemas/PoolId"
                grantedFlavours:
                  type: array
                  items:
                    type: object
                    required:
                      - flavourId
                      - numFlavour
                    properties:
                      flavourId:
                        $ref: "#/components/schemas/FlavourId"
                      numFlavour:
                        type: integer
                        format: int32
                        description: Count of flavour
                  minItems: 1
      responses:
        "204":
          description: Updated Resource reservation status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/partnerDetailsCallbackLink':
    post:
      operationId: PartnerDetailsCallback
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: |
          OP uses this callback api to notify partner OP about change in federation details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederationDetailResponseData'
      responses:
        "204":
          description: Expected response to a successful call back processing
        "400":
          $ref: '#/components/responses/400'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
      security:
        - notifClientCredentials: [fed-mgmt-notif]